requirements.txt
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
celery==4.0.2
crc16==0.1.1
Django==1.10.5
django-filter==1.0.1
django-rest-auth==0.9.0
djangorestframework==3.5.3
kombu==4.0.2
packaging==16.8
pkg-resources==0.0.0
py==1.4.32
pyparsing==2.1.10
pytest==3.0.6
pytest-django==3.1.2
python-redis==0.1.4
pytz==2016.10
redis==2.10.5
six==1.10.0
vine==1.1.3
